+++ pkg/pillar/cmd/vaultmgr/vaultmgr.go
@@ -39,8 +39,8 @@ import (
 	"github.com/lf-edge/eve/pkg/pillar/pidfile"
 	"github.com/lf-edge/eve/pkg/pillar/pubsub"
 	"github.com/lf-edge/eve/pkg/pillar/types"
-	"github.com/lf-edge/eve/pkg/pillar/utils"
 	"github.com/lf-edge/eve/pkg/pillar/utils/persist"
+	"github.com/lf-edge/eve/pkg/pillar/utils/wait"
 	"github.com/lf-edge/eve/pkg/pillar/vault"
 	"github.com/lf-edge/eve/pkg/pillar/zfs"
 	"github.com/sirupsen/logrus"
@@ -202,7 +202,7 @@ func Run(ps *pubsub.PubSub, loggerArg *logrus.Logger, logArg *base.LogObject, ar
 	ps.StillRunning(agentName, warningTime, errorTime)
 
 	// Wait until we have been onboarded aka know our own UUID, but we don't use the UUID
-	err := utils.WaitForOnboarded(ps, log, agentName, warningTime, errorTime)
+	err := wait.WaitForOnboarded(ps, log, agentName, warningTime, errorTime)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -318,7 +318,7 @@ func runInline(ps *pubsub.PubSub, command string, _ []string) int {
 			return 1
 		}
 	case "waitUnsealed":
-		if err := utils.WaitForVault(ps, log, agentName, warningTime, errorTime); err != nil {
+		if err := wait.WaitForVault(ps, log, agentName, warningTime, errorTime); err != nil {
 			log.Fatal(err)
 			return 1
 		}
