+++ pkg/pillar/utils/common.go
@@ -1,0 +1,11 @@
+// Copyright (c) 2024 Zededa, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+package utils
+
+import (
+	uuid "github.com/satori/go.uuid"
+)
+
+// Really a constant
+var nilUUID = uuid.UUID{}
