+++ pkg/pillar/utils/wait/waitfor.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Zededa, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package utils
+package wait
 
 import (
 	"time"
